Chapter 11/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Leonardo Lazzaro <llazzaro@dc.uba.ar>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.7"
psycopg = {extras = ["binary"], version = "^3.1.12"}
django-extensions = "^3.2.3"
django-storages = "^1.14"
django-widget-tweaks = "^1.5.0"
django-redis = "^5.3.0"
pillow = "^10.0.1"
django-ninja = "1.0b2"
pyjwt = "^2.8.0"
django-ratelimit = "^4.1.0"
gunicorn = "^21.2.0"
opentelemetry-instrumentation-django = "^0.42b0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-prometheus = "^1.12.0rc1"
opentelemetry-api = "^1.21.0"
opentelemetry-exporter-jaeger = "^1.21.0"
opentelemetry-exporter-otlp = "^1.21.0"
django-prometheus = "^2.3.1"
django-allow-cidr = "^0.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
django-extensions = "^3.2.3"
pytest-django = "^4.5.2"
factory-boy = "3.0.1"
pytest-bdd = "^7.0.0"
pytest-cov = "^4.1.0"
pytest-xdist = "^3.3.1"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
